Add files via upload
This website is a dog daycare webiste, allowing users to review the services and making bookings.
The website uses an online database for storing data, using php to connect them. the system allows for
users to create accounts with passwords encrypted. To ensure this the blowfish method was used for encrypting passwords